String format problem with musl C library
There are other places in the codebase where time values are
formatted with `%llu` and an explicit `(unsigned long long)`
cast, as printf argument types should not be ambiguous and
always have the expected size.

IssueID #565

+++ src/gmcerts.c
@@ -256,9 +256,9 @@ void serialize_GmCerts(const iGmCerts *d, iStream *trusted, iStream *identsMeta)
         iConstForEach(StringHash, i, d->trusted) {
             const iTrustEntry *trust = value_StringHashNode(i.value);
             format_String(&line,
-                          "%s %ld %s\n",
+                          "%s %llu %s\n",
                           cstr_String(key_StringHashConstIterator(&i)),
-                          integralSeconds_Time(&trust->validUntil),
+                          (unsigned long long) integralSeconds_Time(&trust->validUntil),
                           cstrCollect_String(hexEncode_Block(&trust->fingerprint)));
             write_Stream(trusted, &line.chars);
         }
